Update README.md
Tried to add a new row for unpaved MUPs... it seemed to shift everything down by a row, but created a ton of deletions and insertions to do so... quite a mess to eyeball in the preview, but I think it will look okay. Is there a better way?

Also not confident about how I tried adding a link in the [fine_gravel] tag, so please check that out.