Paring down info to the 'need to know'
I think the notes around physically separated will just confuse people (and they don't need to know this for the sake of tagging - that's about how we'll be using the tags later). Lanes, according to James, are basically done. So I just left the parking and cycle lane stuff as mission critical